+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+2021-04-23.01
+-------------
+
+* Handle `Date` the same way we handle `Timestamp`. In Excel, all dates are
+  Numbers. When the result is a Number, we don't interpret it as a Workbench
+  Date -- same as with timestamp, users see the unformatted Excel Numbers.)
+
 2020-09-15.02
 -------------